Add has_suggested/observed to algorithms
Why:

To avoid passing multiple times the same points to the algorithms,
the producer needs to know whether an algorithm actually observed a
point. Until now the producer was keeping track on of the point passed
to infer which to pass later on, but some algorithms would not observe
points until they had suggested it (ex Hyperband). This caused issue
with the resumption of these algorithms as they would ignore the points
and would never be presented the points again.