Register trials in batch
Why:

It is faster to register trial all in batch and continue suggesting if
some failed than rolling back and restarting each time one trial fails
to register (because is duplicate). With the new sampling scheme there
should be no duplicate locally anyway. Duplicates could only appear when
there is multiple workers, in which case the first trial and all next
would be duplicates and thus the whole batch must be trashed.